+++ lib/l10n/app_localizations_de.dart
@@ -929,6 +929,9 @@ class AppLocalizationsDe extends AppLocalizations {
   String get diskUsage => 'Festplattennutzung';
 
   @override
+  String get diskAvailable => 'verfügbar';
+
+  @override
   String get dnsDomainList => 'DNS-Domänenliste';
 
   @override